feat: useSearchFocus 훅 구현
