+++ src/features/search/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useSearchedKeywordStorage } from './useSearchedKeywordStorage';
 export { default as useDebounce } from './useDebounce';
+export { default as useSearchFocus } from './useSearchFocus';
